Add files via upload
 distributed file system through socket 
programming. 
The distributed file system has three servers: 
 Smain 
 Spdf 
 Stext 
and can support multiple client connections. 
Section A – Servers : Smain, Spdf and Stext
Clients are allowed to upload/store three file types (.c,.pdf and .txt) onto Smain, however 
Smain only stores .c files locally and transfers all .pdf files to the Spdf sever and .txt files to 
the Stext sever (in the background). Clients are not aware of this operation and assume all 
files are stored at Smain. 
All clients communicate with Smain only and are not aware of the presence of Spdf and 
Stext.
Upon receiving a connection request from a client, Smain forks a child process that 
services the client request exclusively in a function called prcclient() and (Smain) 
returns to listening to requests from other clients. 
o The prcclient() function enters an infinite loop waiting for the client to send a 
command 
o Upon the receipt of a command from the client, prclient() performs the action 
required to process the command as per the requirements listed in section B 
and returns the result to the client 
 Spdf and Stext act as servers to Smain and service its requests based on the 
commands entered in client24s

ufile filename destination_path 
Transfers (uploads) filename from the PWD of the client to smain

dfile filename 
Transfers (downloads) filename from Smain to the PWD of the client

rmfile filename 
Removes (deletes) filename from Smain to the PWD of the client

dtar filetype 
Creates a tar file of the specified file type and transfers (downloads) the tar file from 
Smain to the PWD of the client 

display pathname
Transfers (downloads) filename from Smain to the PWD of the client